Created an angular 2 string with watermark component
